Split UserRestController into 3 seperate controllers and externalized the AccessDeniedException handler
