+++ src/main/java/ch/wisv/areafiftylan/controller/UserProfileRestController.java
@@ -1,0 +1,98 @@
+package ch.wisv.areafiftylan.controller;
+
+import ch.wisv.areafiftylan.dto.ProfileDTO;
+import ch.wisv.areafiftylan.model.Profile;
+import ch.wisv.areafiftylan.model.User;
+import ch.wisv.areafiftylan.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+import static ch.wisv.areafiftylan.util.ResponseEntityBuilder.createResponseEntity;
+
+@RestController
+@RequestMapping("/users/current")
+@PreAuthorize("isAuthenticated()")
+public class UserProfileRestController {
+
+    private UserService userService;
+
+    @Autowired
+    UserProfileRestController(UserService userService) {
+        this.userService = userService;
+    }
+
+    /**
+     * Add a profile to a user. An empty profile is created when a user is created, so this method fills the existing
+     * fields
+     *
+     * @param userId The userId of the user to which the profile needs to be added
+     * @param input  A representation of the profile
+     *
+     * @return The user with the new profile
+     */
+    @PreAuthorize("@currentUserServiceImpl.canAccessUser(principal, #userId)")
+    @RequestMapping(value = "/{userId}/profile", method = RequestMethod.POST)
+    public ResponseEntity<?> addProfile(@PathVariable Long userId, @Validated @RequestBody ProfileDTO input) {
+        User user = userService.addProfile(userId, input);
+
+        // Create HttpHeaders to include the location of the newly created profile
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.setLocation(
+                ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}/profile").buildAndExpand(user.getId())
+                        .toUri());
+
+        return createResponseEntity(HttpStatus.CREATED, httpHeaders,
+                "Profile succesfully created at" + httpHeaders.getLocation(), user.getProfile());
+    }
+
+    /**
+     * Add a profile to the current user. An empty profile is created when a user is created, so this method fills the
+     * existing fields
+     *
+     * @param input A representation of the profile
+     *
+     * @return The user with the new profile
+     */
+    @PreAuthorize("isAuthenticated()")
+    @RequestMapping(value = "/current/profile", method = RequestMethod.POST)
+    public ResponseEntity<?> addProfile(@Validated @RequestBody ProfileDTO input, Authentication auth) {
+        return this.addProfile(((User) auth.getPrincipal()).getId(), input);
+    }
+
+    /**
+     * Change the profile fields of the User. Basically the same as the POST request.
+     *
+     * @param userId The userId of the user to which the profile needs to be added
+     * @param input  A representation of the profile
+     *
+     * @return The user with the changed profile
+     */
+    @PreAuthorize("@currentUserServiceImpl.canAccessUser(principal, #userId)")
+    @RequestMapping(value = "/{userId}/profile", method = RequestMethod.PUT)
+    public ResponseEntity<?> changeProfile(@PathVariable Long userId, @Validated @RequestBody ProfileDTO input) {
+        User user = userService.changeProfile(userId, input);
+
+        return createResponseEntity(HttpStatus.OK, "Profile successfully updated", user.getProfile());
+    }
+
+    /**
+     * Resets the profile fields to null. The profile can't actually be deleted as it is a required field.
+     *
+     * @param userId The userId of the user which needs the profile reset
+     *
+     * @return Empty body with StatusCode OK.
+     */
+    @PreAuthorize("hasRole('ADMIN')")
+    @RequestMapping(value = "/{userId}/profile", method = RequestMethod.DELETE)
+    public ResponseEntity<?> resetProfile(@PathVariable Long userId) {
+        Profile profile = userService.resetProfile(userId);
+        return createResponseEntity(HttpStatus.OK, "Profile successfully reset", profile);
+    }
+}
