+++ src/main/java/ch/wisv/areafiftylan/exception/GlobalControllerExceptionHandler.java
@@ -1,0 +1,22 @@
+package ch.wisv.areafiftylan.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import static ch.wisv.areafiftylan.util.ResponseEntityBuilder.createResponseEntity;
+
+/**
+ * This class handles exceptions for ALL Controllers. Whenever a Controller throws an exception that is listed here, the
+ * corresponding method will be used to handle the exception.
+ */
+@ControllerAdvice
+class GlobalControllerExceptionHandler {
+
+    @ExceptionHandler(AccessDeniedException.class)
+    public ResponseEntity<?> handleAccessDeniedException(AccessDeniedException ex) {
+        return createResponseEntity(HttpStatus.FORBIDDEN, "Access denied");
+    }
+}
